requirements.txt
streamlit==1.23.1
docx2txt==0.8
requests==2.31.0
python-dotenv==1.0.0
python-docx==0.8.11